% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\name{addTicker}
\alias{addTicker}
\title{addTicker}
\usage{
addTicker(
  name,
  yahoo_name,
  trading_class,
  multiplier = 100,
  type = "STK",
  currency = "USD",
  exchange = "SMART",
  opt_exchange = "SMART",
  IV = "YES",
  expiration = ""
)
}
\arguments{
\item{name}{IBKR security name}

\item{yahoo_name}{Yahoo finances download security name.  If not provided, will be equal to \code{name}.}

\item{trading_class}{chain trading class that will be used to retrieve option data from IBKR. If not provided, will be equal to \code{name}.}

\item{multiplier}{chain multiplier, default is 100}

\item{type}{IBKR security type, default is STK (stock). Could be also FUT (future), IND (Index),...}

\item{currency}{IBKR currency value, i.e. USD, EUR, CHF,... Default is USD}

\item{exchange}{exchange name where security price can be retrieved, default is SMART}

\item{opt_exchange}{exchange name where option price for the security can be retrieved, default is SMART}

\item{IV}{String character - whether IV should be looked at or not , default is YES}

\item{expiration}{String character - should be a date for YYYYMM for futures or YYYYMMDD for bills, bonds. Default is empty string}
}
\value{
record added and display number of records added, i.e. 1 (normal case) or 0 (no line deleted, error case)
}
\description{
This function is used by scanner functions to add one ticker in DB
}
\examples{
\dontrun{
addTicker("SPY")
}
}
