% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_f.R
\name{as.character.ldtf}
\alias{as.character.ldtf}
\title{Convert Frequency to Character}
\usage{
\method{as.character}{ldtf}(x, ...)
}
\arguments{
\item{x}{The value of the frequency, which must be an \code{ldtf} object returned from the \code{f.?} functions.}

\item{...}{Additional arguments.}
}
\value{
A string representation of the value of the frequency.
}
\description{
This function converts a frequency to its string representation. The format is explained in the \code{f.?} functions.
}
