% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_f.R
\name{as.frequency}
\alias{as.frequency}
\title{Converts Character to frequency}
\usage{
as.frequency(str, classId)
}
\arguments{
\item{str}{value of the frequency as a valid character. You might have get it from \code{\link{as.character.ldtf}} function}

\item{classId}{class id of the frequency as a valid character.}
}
\value{
An object of class 'ldtf'
}
\description{
Use this method to convert back your character to a frequency. You need the class id information.
}
