% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{as.numeric.ldtv}
\alias{as.numeric.ldtv}
\title{Coerce Variable to 'numeric'}
\usage{
\method{as.numeric}{ldtv}(x, ...)
}
\arguments{
\item{x}{Variable with \code{data} field.}

\item{...}{Other arguments.}
}
\value{
\code{data} in x.
}
\description{
Coerce Variable to 'numeric'
}
