% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sousligne.R
\name{build_sub_line}
\alias{build_sub_line}
\title{build_sub_line function}
\usage{
build_sub_line(
  type = "",
  sym = "",
  strike = 0,
  expdate = "",
  DTE = NA,
  price = 0,
  u_price = 0,
  IV = 0,
  mul = NA,
  delta = NA,
  deltanotional = NA,
  gamma = NA,
  theta = NA,
  vega = NA
)
}
\arguments{
\item{type}{Put Call or Stock}

\item{sym}{ticker name Empty default value}

\item{strike}{strike value - 0 is default}

\item{expdate}{expiration date Empty default value}

\item{DTE}{number of days to expiration -NA is default}

\item{price}{initial option price - 0 is default}

\item{u_price}{underlying price - 0 is default}

\item{IV}{implied vol - 0 is default}

\item{mul}{multiplier - NA is default}

\item{delta}{delta value  - NA is default}

\item{deltanotional}{delta notional value (delta dollars)  - NA is default}

\item{gamma}{gamma value  - NA is default}

\item{theta}{theta value  - NA is default}

\item{vega}{vega value  - NA is default}
}
\value{
a data frame with the following arguments:
\itemize{
\item \code{instrument},
\item \code{type},
\item \code{strike},
\item \code{mul},
\item \code{price},
\item \code{IV},
\item \code{DTE},
\item \code{delta},
\item \code{deltanotional}
\item \code{gamma},
\item \code{theta},
\item \code{vega},
}
}
\description{
This utility transforms a set of arguments into a dataframe. Any missing argument gets a default value,
}
\examples{
build_sub_line()
build_sub_line(sym="STOCK",type="Stock",price=100,
mul=1,delta=5,deltanotional = 5*100)
build_sub_line(sym="STOCK",type="Put",strike=100,expdate="20240119",
price=1.5,mul=100,delta=50,deltanotional=50*100,
vega=-24,theta=12)
}
