% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{c_to_usd}
\alias{c_to_usd}
\title{c_to_usd}
\usage{
c_to_usd(amount, currency)
}
\arguments{
\item{amount, currency}{amount is the number to be converted, currency is a string whose value is either EUR, CHF}
}
\description{
This function converts the amount of currency into USD, using currency pairs values stored in DB
}
\details{
It merely performs a multiplication of the amount by currency pair value, using getStoredValue.
This function can be vectorized
}
\examples{
c_to_usd(100.45,"EUR")
c_to_usd(c(10000,500),c("CHF","EUR"))
c_to_usd(c(750.543,10),c("USD","EUR"))
c_to_usd(c(500,10),c("CAD","EUR"))
}
\keyword{currency}
\keyword{trading}
