% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{change}
\alias{change}
\title{change}
\usage{
change(x, y)
}
\arguments{
\item{x}{is a vector of integers, cannot include NA}

\item{y}{is a vector of integers, cannot include NA#}
}
\value{
numeric number
}
\description{
Computes the relative change between x and y, x is start value, y end value.
}
\details{
x and y may be negative, but if x = 0, then change returns NA (float)
x and y may be vectors of numbers
}
\examples{
change(1,1.1)  ### Should be equal to 0.1
change(1,1)  ### Should be equal to 0
change(-1,-1.1)  ### Should be equal to -0.1
change(0,1) ## NA
change(-1,-0.9)  ### Should be equal to 0.1
}
