% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{create_centered_plot}
\alias{create_centered_plot}
\title{creates a graph centered on median values for x and y series}
\usage{
create_centered_plot(
  data,
  symbol_col,
  x_col,
  y_col,
  title = "IMplied vol Dashboard"
)
}
\arguments{
\item{data}{data frame with data to show}

\item{symbol_col}{column name containing symbols}

\item{x_col}{x column name}

\item{y_col}{y column name}

\item{title}{string - graph title}
}
\value{
ggplot2 object
}
\description{
creates a graph centered on median values for x and y series
}
\examples{
data <- data.frame(
  Symbole = c("A", "B", "C", "D", "E"),
  IVP = c(0.01, 0.02, 0.03, 0.04, 0.05),
  `IV-1month` = c(0.05, 0.04, 0.03, 0.02, 0.01)
)
create_centered_plot(data, "Symbole", "IVP", "IV-1month", "Exemple de graphique")
}
