% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{currency_convert}
\alias{currency_convert}
\title{currency_convert}
\usage{
currency_convert(amount, currency)
}
\arguments{
\item{amount, currency}{amount is the number to be converted, currency is a string whose value is either EUR, CHF}
}
\description{
This function converts the amount of currency into USD, using getCurrencyPairs
}
\details{
First it retrieves today's currency pairs by calling getCurrencyPairs() function.

Then it just performs a multiplication of the amount by currency pair value. This function can be vectorized
}
\keyword{Examples}
\keyword{currency}
\keyword{currency_convert(100.45,"EUR")}
\keyword{currency_convert(c(10000,500),c("CHF","EUR"))}
\keyword{currency_convert(c(750.543,10),c("USD","EUR"))}
\keyword{trading}
