% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{currency_format}
\alias{currency_format}
\title{currency_format}
\usage{
currency_format(amount, currency)
}
\arguments{
\item{amount, currency}{amount is the number to be displayed, currency is a string whose value is either EUR, CHF or USD}
}
\description{
This function returns a string with amount and currency symbol
}
\details{
It defines local labeling functions for CHF, EUR and USD.

Currency symbol (for EUR and USD) are also taken into consideration.

Only 2 digits after decimal are displayed. Big mark (3 digits separator) is empty.
}
\examples{
currency_format(100.45,"EUR")
currency_format(10000,"CHF")
currency_format(758.458,"USD")
currency_format(100000.455,"EUR")
currency_format(c(100,40),c("EUR","USD"))
}
\keyword{currency}
\keyword{trading}
