% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{currency_sign}
\alias{currency_sign}
\title{currency_sign}
\usage{
currency_sign(currency)
}
\arguments{
\item{currency}{a string or a vector of strings - possible values are EUR, CHF,...}
}
\value{
a character or a vector of characters
}
\description{
This function returns display sign for a given currency
}
\details{
It looks into Currencies table from DB and then returns corresponding currency sign in table.
}
\examples{
currency_sign("USD")
currency_sign("EUR")
currency_sign("CAD")
currency_sign(c("CHF", "USD"))
}
