% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data.pcp}
\alias{data.pcp}
\title{Load PCP Dataset}
\usage{
data.pcp(dirPath, makeReal = FALSE)
}
\arguments{
\item{dirPath}{A character string representing the path to the downloaded data directory.}

\item{makeReal}{A logical value indicating whether to convert nominal variables to real using the first column (which must be US-CPI). If \code{TRUE}, \code{dirPath} must also contain a file with the US CPI.}
}
\value{
A list with data, descriptions, etc.
}
\description{
Use this function to load the 'IMF's Primary Commodity Prices' dataset and create a Date-Series table.
}
