% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{deleteJournalEntry}
\alias{deleteJournalEntry}
\title{deleteJournalEntry}
\usage{
deleteJournalEntry(entryId)
}
\arguments{
\item{entryId, }{integer - entry key for the entry to be removed}
}
\value{
number of lines deleted, if no error it returns 1. It returns 0 if there is an error
}
\description{
This function removes an entry from Journal table.
}
\details{
No verification is done prior sending request to DB, i.e. if entryId exists corresponding line is removed otherwise it returns 0
}
