% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{enter_numerical_data}
\alias{enter_numerical_data}
\title{Enter numerical data}
\usage{
enter_numerical_data(sym)
}
\arguments{
\item{sym}{This is the name data to be filled in by end user}
}
\value{
numerical input by end user, if cannot be coerced to a number then NA
}
\description{
This very simple code piece is intended to enter one numerical value and then returns it.
This is to be used whenever additional data must be entered that was not initially foreseen to be entered
(e.g. due to lack of IBKR connection).
}
\details{
There are 2 cases:
\itemize{
\item R is in interactive mode (e.g. from console)
\item R is called from batch.
These 2 sub-cases are handled with different function calls (\code{readline} vs \code{readLines})
}
}
