% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{f.list.string}
\alias{f.list.string}
\title{Create a \code{List-String} Frequency}
\usage{
f.list.string(items, value)
}
\arguments{
\item{items}{The items in the list.}

\item{value}{The current item.}
}
\value{
An object of class \code{ldtf}, which is also a list with the following members:
\item{class}{Determines the class of this frequency.}
\item{items}{Determines the \code{items}.}
\item{value}{Determines the \code{value}.}
}
\description{
This frequency is typically used for labeled data. It is generally a list, but it can also be used to label observations outside this list.
}
\details{
In order to use the \code{\link{as.frequency}} function for this type of frequency,
you need the following information:
\itemize{
\item \strong{Character Format}: \code{"..."} (where '...' represents the \code{value})
\item \strong{Class Id}: \code{Ls} or \code{Ls:...} (where '...' represents the semi-colon-separated \code{items})
}
}
\examples{

L0 <- f.list.string(c("A","B","C","D"), "C")

L0_value_str <-  as.character(L0) # this will be 'C'.
L0_class_str <- get.class.id(L0) # this will be 'Ls:A;B;C;D'.

L_new <- as.frequency("A", "Ls:A;B;C;D")
L_new0 <- as.frequency("A", "Ls") # compared to the previous one, its items will be empty

# Don't make the following mistakes:
\donttest{
L_invalid <- try(as.frequency("E", "Ls:A;B;C;D")) # 'E' is not a member of the list
L_invalid <- try(f.list.string(c("A","B","C","D"), "E"))
}

}
