% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{f.quarterly}
\alias{f.quarterly}
\title{Creates a Quarterly Frequency}
\usage{
f.quarterly(year, quarter)
}
\arguments{
\item{year}{Year of the observation.}

\item{quarter}{Quarter of the observation (It should be between 1 and 4).}
}
\value{
An object of class 'ldtf'. It is also a list with the following members:
\tabular{ll}{
\code{class} \tab Determines the class of this frequency.\cr
\code{year} \tab Determines the \code{year}.\cr
\code{quarter} \tab Determines the \code{quareter}.
}
}
\description{
Use it to create a frequency for time-series data that occurs quarterly.
}
\details{
In order to use \code{\link{as.frequency}} function for this type of frequency,
you need the following information:
\itemize{
\item \strong{Character Format} \code{"#q#"} (first '#' is the \code{year}, second '#' is the
\code{quarter}; e.g., 2010q3 or 2010q4. Note that 2000q0 or 2000q5 are invalid.
\item \strong{Class Id} \code{"q"}
}
}
\examples{

q0 <- f.quarterly(2020, 2)
#     this is a quarterly frequency that refers to the second quarter of the year 2021.

q0_value_str <-  as.character(q0) # this will be '2020Q2'.
q0_class_str <- get.class.id(q0) # this will be 'q'.

q_new <- as.frequency("2021q3", "q")
#      this is a quarterly frequency that refers to the third quarter of the year 2021.

# Don't make the following mistakes:
\donttest{
q_invalid <- try(f.quarterly(2020, 0))
q_invalid <- try(f.quarterly(2020, 5))
q_invalid <- try(as.frequency("2021q0", "q"))
q_invalid <- try(as.frequency("2021q5", "q"))
q_invalid <- try(as.frequency("2021", "q"))
}

}
