% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get.descriptive}
\alias{get.descriptive}
\title{Get Descriptive Statistics}
\usage{
get.descriptive(data, type, skipNAN)
}
\arguments{
\item{data}{A numeric vector that contains the data.}

\item{type}{Its a character array that determines the type of the descriptive statistics.}

\item{skipNAN}{If \code{TRUE}, it checks for and skips any \code{NA}s. If you are sure that data does not have \code{NA} set it to be \code{False}.}
}
\value{
An array with different sizes based on the inputs
}
\description{
Use this function to get descriptive statistics of a numeric vector.
}
