% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatility.R
\name{get180dIV}
\alias{get180dIV}
\title{get180dIV}
\usage{
get180dIV(tickers)
}
\arguments{
\item{tickers}{data frame with each row a ticker}
}
\value{
a tibble with Name and IV180 as columns. IV180 is rounded to 3rd decimal
}
\description{
Get 180 days IV for tickers
}
\details{
Computes 180 days IV using option IVs for 180 days target days expiration
}
\examples{
get180dIV(getTicker("AI"))
get180dIV(getTickers(c("AI", "SPX")))
}
