% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatility.R
\name{get30dIV}
\alias{get30dIV}
\title{get30dIV}
\usage{
get30dIV(tickers)
}
\arguments{
\item{tickers}{data frame with each row a ticker}
}
\value{
a tibble with Name and IV30 as columns. IV30 is rounded to 3rd decimal
}
\description{
Get 30 days IV for tickers.
}
\details{
Computes 30 days IV using option IVs for 30 days target expiration
}
\examples{
get30dIV(getTicker("AI"))
get30dIV(getTickers(c("AI", "SPX")))
}
