% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_rate_utils.R
\name{get_chf_rates}
\alias{get_chf_rates}
\title{Get CHF interest rates from NASDAQ Data Link}
\usage{
get_chf_rates(verbose = FALSE)
}
\value{
A data frame with CHF interest rates
}
\description{
Fixed function for getting CHF rates with environment variable authentication
}
\keyword{internal}
