% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_rate_utils.R
\name{get_de_yield}
\alias{get_de_yield}
\title{get_de_yield}
\usage{
get_de_yield()
}
\value{
data frame with record date and 2 years EUR value
}
\description{
Get German 2Y yields from FRED
}
\details{
Not used because it does not get updated often enough - 2025-03-01, 2.746\% although current rate is 1.688\% on May 1st, 2025
}
\keyword{internal}
