% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_rate_utils.R
\name{get_swiss_bond_yields_numeric}
\alias{get_swiss_bond_yields_numeric}
\title{Get just the numeric Swiss bond yields}
\usage{
get_swiss_bond_yields_numeric(
  maturities = c("1Y", "2Y"),
  use_static = FALSE,
  debug = FALSE
)
}
\arguments{
\item{maturities}{Character vector of desired maturities.}

\item{use_static}{Logical. If TRUE, uses static values}

\item{debug}{Logical. If TRUE, outputs debugging information}
}
\value{
A named numeric vector with yields
}
\description{
Get just the numeric Swiss bond yields
}
\keyword{internal}
