% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_rate_utils.R
\name{get_usd_rates}
\alias{get_usd_rates}
\title{Get USD interest rates from FRED}
\usage{
get_usd_rates()
}
\value{
A data frame with USD interest rates
}
\description{
Get USD interest rates from FRED
}
\keyword{internal}
