% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{getAccountLive}
\alias{getAccountLive}
\title{getAccountLive}
\usage{
getAccountLive()
}
\value{
No value
}
\description{
This function reads last portfolio from Gonet and then deduces account record similar to IBKR
and stores it into DB "Account" table.
The tricky piece is to manage Cash positions
}
\examples{
\dontrun{
getAccountLive()
}
}
