% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getActiveTrades}
\alias{getActiveTrades}
\title{getActiveTrades}
\usage{
getActiveTrades(account)
}
\arguments{
\item{account}{string, account name}
}
\value{
All trades related to active position stored in Trades table from mydb DB. Format is the following:
\code{TradeNr, Account, TradeDate, Strategy, Instrument, Ssjacent, Pos, Prix,
Comm., Total, Exp.Date, Risk, Reward, PnL, Statut, Currency}
}
\description{
This function works only for IBKR accounts not for Gonet account.
It reads active trades from given selected account and returns all corresponding records from table Trades in DB.
}
\details{
It retrieves all trades whose \code{Statut} is different from closed. It also verifies that \code{TradeNr,TradeDate,Pos,Prix, Comm., Total, Risk, Reward, PnL} are all numeric,
and if not, displays an error message and converts them
}
\examples{
\dontrun{
getActiveTrades("DUxxx")
}
}
