% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_rate_utils.R
\name{getAllTenors}
\alias{getAllTenors}
\title{getAllTenors}
\usage{
getAllTenors(currency = "All")
}
\arguments{
\item{currency}{String. Currency code like "EUR" or "USD", or "All" to get all currencies tenors."All" is default value.}
}
\value{
data frame from DB: it contains \code{Name, last_ir_update, ir1week, ir1month, ...}
}
\description{
Get all tenors, interest rate from database for a given currency
}
\details{
This function will retrieve all tenors stored in DB for a given currency.

A warning message will be displayed if data stored in DB is older than 8 days.
}
\examples{
\dontrun{
getAllTenors("EUR")
getAllTenors("USD")
getAllTenors()
}
}
