% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getBjSPrice}
\alias{getBjSPrice}
\title{getBjSPrice}
\usage{
getBjSPrice(
  type,
  S,
  K,
  r = as.numeric(config::get("interest_rate")),
  DTE,
  sig,
  mul,
  div = 0
)
}
\arguments{
\item{type}{a string, if equal to "Put" then getBjSPutPrice is called, if equal to "Call" getBjSCallPrice is called, if equal to "Stock" returns underlying price, else returns NA}

\item{S}{Spot price or current underlying price}

\item{K}{Strike price}

\item{r}{Interest rate - default value is defined in config.yml file}

\item{DTE}{Number of days to expiration (may be a decimal number - useful as expiration may happen at 4:00pm or in the morning (index case))}

\item{sig}{annualized volatility}

\item{mul}{contract multiplier - generally 100 - but equal to 10 for ESTX OESX option chains}

\item{div}{annualized dividend yield (i.e. sum of present values of upcoming dividends over 1 year divided by current spot price). Default value is 0.}
}
\description{
This function supplies a contract price, i.e. either an option price x multiplier
or a stock price. The option price is based upon binomial tree model or a stock price.
}
\details{
This function can be vectorized. It handles also special cases not handled correctly by derivmkts, such as:
\itemize{
\item DTE=0 and S=K -> option price = 0
\item DTE < 0 -> option price = 0
\item S=K=0 -> option price = 0
}
}
\examples{
getBjSPrice(type="Put",S=100,K=100,r=0.04,DTE=5,sig=0.25,mul=100)
getBjSPrice(type="Call",S=22,K=22.5,r=0.04,DTE=5,sig=0.3,mul=100,div=0.4)
getBjSPrice(type="Stock",S=100,K=NA,DTE=NA,sig=NA,mul=1)
}
