% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getBjSPutPrice}
\alias{getBjSPutPrice}
\title{getBjSPutPrice}
\usage{
getBjSPutPrice(
  S,
  K,
  r = as.numeric(config::get("interest_rate")),
  DTE,
  sig,
  div = 0
)
}
\arguments{
\item{S}{Spot price or current underlying price}

\item{K}{Strike price}

\item{r}{Interest rate - default value is defined in config.yml file}

\item{DTE}{Number of days to expiration (may be a decimal number - useful as expiration may happen at 4:00pm or in the morning (index case))}

\item{sig}{annualized volatility}

\item{div}{annualized dividend yield (i.e. sum of present values of upcoming dividends over 1 year divided by current spot price). Default value is 0.}
}
\description{
This function provides a Bjerksund Stensland approximation of option put price by iterating 100 times, assuming that current price is risk neutral price.
}
\details{
Approximation is done using a binomial tree. There is a given probability of going up or down, using an up multiplier of 1.5 and down multiplier of 0.5
and having a risk neutral hypothesis - i.e. the current price does not allow any bias towards high or low option price
It is assumed that option are American style
If dividend exists, option may therefore be exercised.

Notice that results are rounded to 3 digits after decimal point.
}
\examples{
getBjSPutPrice(S=100,K=100,r=0.04,DTE=5,sig=0.25,div=0)
getBjSPutPrice(S=22,K=22.5,r=0.04,DTE=5,sig=0.3,div=0)
}
