% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getClosedTrades}
\alias{getClosedTrades}
\title{getClosedTrades}
\usage{
getClosedTrades(account, windowDate = Sys.Date() - 200)
}
\arguments{
\item{account}{string account selection, either Live or Simu}

\item{windowDate}{Integer, format is YYYYmmdd. Only records whose original date is greater than window date will be returned. Default value is today minus 200 days.}
}
\value{
All trades related to closed position stored in Trades table from mydb DB. Format is the following:
\code{TradeNr, Account, TradeDate, Strategy, Instrument, Ssjacent, Pos, Prix,
Comm., Total, Exp.Date, Risk, Reward, PnL, Statut, Currency}
}
\description{
This function works only for IBKR accounts not for Gonet account.
It retrieves closed trades from given selected account, after a given window date, and returns all corresponding records from table Trades in DB.
}
\details{
It retrieves all trades whose \code{Statut} is equal to closed.
It also verifies that \code{TradeNr, TradeDate, Pos, Prix, Comm., Total, Risk, Reward, PnL} are all numeric,
and if not, displays an error message and converts them
}
