% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{getCurrencyPairs}
\alias{getCurrencyPairs}
\title{getCurrencyPairs}
\usage{
getCurrencyPairs()
}
\description{
This function provides a pair EUR/USD, CHF/USD.
}
\details{
It looks into CurrencyPairs table and requests
the last record. It does not try to retrieve more up to date value from IBKR or other sources.
See \code{getIBKR()} to get more up to date data from IBKR
}
