% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividend_utils.R
\name{getDividendYield}
\alias{getDividendYield}
\title{getDividendYield}
\usage{
getDividendYield(tickers, verbose = FALSE)
}
\arguments{
\item{tickers}{data frame of tickers}
}
\value{
a list of double, -1 for tickers where no value could be computed, even with Yahoo
}
\description{
Returns dividend yield for one or a vector of symbols, looking at IBKR or Yahoo through internal getSingleDivYield function.
}
