% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getDummy}
\alias{getDummy}
\title{private method to generate dummy variable}
\usage{
getDummy(table, colName, pre = "", min_unique_skip = 6, uniques = NULL)
}
\arguments{
\item{table}{data}

\item{colName}{categorical column}

\item{pre}{a string to put before the name of the variables}

\item{min_unique_skip}{if number of unique values is equal or larger, it returns \code{NULL}}

\item{uniques}{if not \code{NULL}, it skips finding unique values and uses
the given list. Also, if \code{colName} column is missing,
it creates zero variables for the given items}
}
\value{
\code{data} (list of dummy variables) and \code{uniques} (unique values)
}
\description{
private method to generate dummy variable
}
