% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{getGonet}
\alias{getGonet}
\title{getGonet}
\usage{
getGonet()
}
\value{
No value
}
\description{
This function retrieves trades from GonetTrades.CSV and deduces current Gonet portfolio positions.
It he retrieves prices from IBKR and then store updated Gonet portfolio positions into DB "Gonet" table.
}
\examples{
\dontrun{
getGonet()
}
}
