% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{getIBKRActiveCurrencyValues}
\alias{getIBKRActiveCurrencyValues}
\title{getIBKRActiveCurrencies}
\usage{
getIBKRActiveCurrencyValues()
}
\value{
No value
}
\description{
This function retrieves active currencies pairs values from DB ActiveCurrencies table, and then requests value from IBKR.
Once data retrieved, it is then stored into DB.
}
\details{
This does not request any value from end-user yet, in case IBKR does not return any value.
}
\examples{
\dontrun{
getIBKRActiveCurrencyValues()
}
}
