% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibkr.R
\name{getIBKRMetrics}
\alias{getIBKRMetrics}
\title{getIBKRMetrics}
\usage{
getIBKRMetrics(sym, reqType = 2, close = FALSE, verbose = FALSE, numcores = 3)
}
\arguments{
\item{sym}{string or vector of strings- IBKR style of ticker, if unknown then function returns -1}

\item{reqType}{integer - should be either 2 or 4, default is 2}

\item{close}{Boolean TRUE/FALSE, default is FALSE. if TRUE then retrieve last close price from IBKR, if FALSE then retrieve market price from IBKR}

\item{verbose}{Boolean TRUE/FALSE, default is FALSE. if TRUE then print result of retrieved data in IBKR}
}
\value{
a data frame with the following fields: \code{datetime} (formatted date and time), \code{sym} (ticker name),
\code{price} (price as double),  \code{iv180} IV for 6 months expiration, \code{iv30} IV for 1 month expiration,
\code{ivp} IV percentile, \code{rv30}, realized volatility for the last month,  \code{rvp} realized volatility percentile.
}
\description{
Retrieves a price or a list of price from IBKR and
returns a data frame with date and time (formatted), the list of tickers and corresponding prices, equal to Nan if no price found.
Additionally it stores prices in DB Prices table (only if price different from NaN and not a close price)
}
\details{
For a given ticker or a list of tickers, this function returns prices from IBKR, using also Tickers table from DB within Python code called.
If IBKR service is not available, it will return an error code and display an error message: 0 if no IBKR service, -1 if contract is unknown.
if unknown then function returns -1
}
\examples{
\dontrun{
getIBKRMetrics("SPY")
getIBKRMetrics(c("ESTX50", "USO", "ABBN"))
getIBKRMetrics("USO", close=TRUE)
}
}
