% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getImpliedIROpt}
\alias{getImpliedIROpt}
\title{getImpliedIROpt}
\usage{
getImpliedIROpt(type, S, K, sig, DTE, div = 0, price)
}
\arguments{
\item{type}{a string, if equal to "Put" then getBSPutPrice is called, if equal to "Call" getBSCallPrice is called, if equal to "Stock" returns 1, else returns NA}

\item{S}{Spot price or current underlying price}

\item{K}{Strike price}

\item{sig}{Implied Volatility}

\item{DTE}{Number of days to expiration (may be a decimal number - useful as expiration may happen at 4:00pm or in the morning (index case))}

\item{div}{annualized dividend yield (i.e. sum of present values of upcoming dividends over 1 year divided by current spot price). Default value is 0.}

\item{price}{option price}
}
\description{
This function computes the implied interest rate of a given price, based upon Black-Scholes-Merton model
}
\details{
It performs the computation using dichotomia algorithm.
As such this requires that the price function is monotonous. Otherwise a brute force algorithm must be used.

This function cannot be vectorized.
}
\examples{
getImpliedIROpt("Call",S=100,K=100,DTE=5,sig=0.25,div=0,price=1.19455)
getImpliedIROpt(type="Put",S=22,K=22.5,sig=0.4,DTE=5,price=1.1)
}
\keyword{Black-Scholes}
\keyword{trading}
\keyword{volatility}
