% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatility.R
\name{getIV_DTE}
\alias{getIV_DTE}
\title{getIV_DTE
Get IV for one ticker}
\usage{
getIV_DTE(ticker, DTE = 30)
}
\arguments{
\item{ticker}{one ticker - data frame with one row, or list}

\item{dte}{numeric - should not be smaller than 10 days as model would not work correctly for target days smaller than 8 days,
should be less than 2 years (730 days)}
}
\value{
a list containing the following: near_forward, next_forward, near_atm_vol, next_atm_vol
near_variance, next_variance, weights, variance_day, v - v is the implied volatility
}
\description{
Computes IV using option IVs for a defined target days.
It will look for nearest expiration dates compared with DTE for the ticker and then retrieve option prices and IV from IBKR.
}
\examples{
getIV_DTE(getTicker("AI"), 30)
getIV_DTE(getTickers(c("AI", "SPX")), 30)
}
