% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{getLastAdjustedPrice}
\alias{getLastAdjustedPrice}
\title{getLastAdjustedPrice}
\usage{
getLastAdjustedPrice(ticker)
}
\arguments{
\item{ticker}{ticker name, as known by IBKR - can be one name or a vector of names}
}
\value{
a list of values (rounded to 2 decimals) corresponding to last values of tickers
}
\description{
This function takes one ticker (or a vector of tickers) as input and returns the last available adjusted value from Yahoo service.
}
\details{
It calls \code{getLastSymPrice} and returns value column from this data frame.
If ticker is NA or NULL or equal to All or STOCK, it returns NA also
}
\examples{
getLastAdjustedPrice("SPY")
}
