% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{getLastCurrencyPairs}
\alias{getLastCurrencyPairs}
\title{getLastCurrencyPairs}
\usage{
getLastCurrencyPairs()
}
\description{
This function returns pairs EUR/USD, CHF/USD, CAD/USD and tries to get from Yahoo the latest available pair.
}
\details{
It looks into CurrencyPairs table and retrieves last available pairs from Yahoo. If Yahoo pairs are more recent,
it will update CurrencyPairs table in DB and return it. Otherwise it just returns
the last record from DB.

It does not try to retrieve more up to date value from IBKR.
See \code{getIBKR()} to get more up to date data from IBKR
}
