% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_rate_utils.R
\name{getLastRate}
\alias{getLastRate}
\title{Get interest rate from database}
\usage{
getLastRate(currency, DTE = 30)
}
\arguments{
\item{currency}{Currency code}

\item{DTE}{Interest rate target date, default is 30 days}
}
\value{
Interest rate as numeric (decimal, not percent terms) or NA if not available
}
\description{
This function will compare DTE with the different tenors stored in DB.
}
\details{
A warning message will be displayed if data stored in DB is older than 8 days.
}
\examples{
\dontrun{
getLastRate("EUR", 180)
getLastRate("USD")
}
}
