% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{getLastUSDValue}
\alias{getLastUSDValue}
\title{getLastUSDValue}
\usage{
getLastUSDValue(currency)
}
\arguments{
\item{currency}{string - possible values are EUR, CHF,...}
}
\value{
a data frame with \code{date, usd_value} fields,
where \code{date} is an integer format of YYYYMMDD and \code{usd_value} a numeric.
}
\description{
This function returns converted value of a given currency to USD
by trying to get from Yahoo the latest available value and by default
returning last value available in DB. This function cannot be vectorized.
}
\details{
It looks into ConvertToUSD table and then retrieves last available value from Yahoo.
If Yahoo value is different from NA and if it is are more recent than stored value, then
it will update ConvertToUSD table in DB. It returns the stored value anyhow.

It does not try to retrieve more up to date value from IBKR.
See \code{getIBKR()} to get more up to date data from IBKR
}
\examples{
getLastUSDValue("USD")
getLastUSDValue("EUR")
getLastUSDValue("CAD")
getLastUSDValue("CHF")
}
