% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatility.R
\name{getOptionPrices}
\alias{getOptionPrices}
\title{getOptionPrices}
\usage{
getOptionPrices(sym, strikes, expiration)
}
\arguments{
\item{sym}{string - IBKR style of ticker, if unknown then function returns -1}

\item{expiration}{number, date or string - expiration date, format is Y/M/D}
}
\value{
a dataframe with with columns strikes, call, put, empty if no price are found
}
\description{
Retrieves a dataframe of option price from IBKR. This function is not vectorized.
}
\details{
For a given contract and expiration, this function returns a set of prices from IBKR
If IBKR service is not available, or option price not available, or contract does not exist, it will return an error code -1.
}
\examples{
\dontrun{
}
}
