% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getRnR}
\alias{getRnR}
\title{getRnR}
\usage{
getRnR(trade_nr)
}
\arguments{
\item{trade_nr}{an integer or a vector of integers}
}
\value{
a dataframe with column trade number, reward and risk.
Each row is associated with a trade number.
Reward column is the sum of rewards (non-NA) for all given instruments that belong to the trade number
Risk column is the sum of risks (non-NA) for all given instruments that belong to the trade number
}
\description{
This function retrieves a data frame listing reward and risk for a given list of trade numbers
}
\details{
It works by matching all trade numbers in Trades DB against the argument \code{trade_nr} given.
Then it groups by trade number and compute for each trade number the total of reward and risks.
Trades may be opened, adjusted or closed.
If the same trade number is present multiple times, then risk and reward data will be repeated
This is useful for mass getRnR calls
}
