% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{getStoredUSDValue}
\alias{getStoredUSDValue}
\title{getStoredUSDValue}
\usage{
getStoredUSDValue(currency)
}
\arguments{
\item{currency}{string - possible values are EUR, CHF, USD...}
}
\value{
a data frame with \code{date, usd_value} fields,
where \code{date} is an integer format of YYYYMMDD and \code{usd_value} a numeric.
}
\description{
This function retrieves last USD value of a given currency from DB. This function can be vectorized.
}
\details{
It looks into ConvertToUSD table and requests
the last record. It does not try to retrieve more up to date value from IBKR or other sources.
See \code{getIBKR()} to get more up to date data from IBKR
}
\examples{
getStoredUSDValue("USD")
getStoredUSDValue("EUR")
getStoredUSDValue("CAD")
getStoredUSDValue("CHF")
}
