% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{getStrategies}
\alias{getStrategies}
\title{getStrategies}
\usage{
getStrategies()
}
\value{
A vector of characters sorted by alphabetical order, equal to the list of current strategies
}
\description{
This function is used by RReporting functions and allows to obtain all valid strategies from DB
}
