% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{getSymIntervalDate}
\alias{getSymIntervalDate}
\title{getSymIntervalDate}
\usage{
getSymIntervalDate(sym, from_date, to_date = Sys.Date())
}
\arguments{
\item{sym}{one or a vector of symbols}

\item{from_date}{a start date from which to retrieve symbols}

\item{to_date}{a end date to which to retrieve symbols - Default is today}
}
\value{
Data frame with columns: date, ticker, Open, High, Low, Close, Adjusted, and Volume
}
\description{
This function gets from Yahoo service all values (Open, High, Low, Close, Volume and Adjusted)
for one of a vector of symbols.
}
\details{
It is based upon getYahooDataRobust, based upon quantmod getSymbols function.
It converts IBKR-style tickers into Yahoo-style tickers first, by looking up into Tickers table.
If not found in Tickers table, then just returns what has been given in \code{sym} argument.
}
\keyword{Yahoo}
