% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{getSymMetricIntervalDate}
\alias{getSymMetricIntervalDate}
\title{getSymMetricIntervalDate}
\usage{
getSymMetricIntervalDate(
  sym,
  from_date,
  to_date = Sys.Date(),
  metric = "Adjusted"
)
}
\arguments{
\item{sym}{one or a vector of symbols}

\item{from_date}{a start date from which to retrieve symbols}

\item{to_date}{a end date to which to retrieve symbols - Default is today}

\item{metric}{column to retrieve from Yahoo service, can be any of: "Open" "High" "Low" "Close" "Volume" "Adjusted". Default value is "Adjusted".}
}
\value{
a tibble: columns are date, symbol names.
Each (non date) column lists \code{metric} values for a given symbol and all dates between  \code{from_date} till \code{to_date}.
}
\description{
This function retrieves a metric from \code{from_date} till \code{to_date} for one or a vector of symbols
This function calls \code{getSymFromDate} with one or a vector of tickers.
}
\examples{
getSymMetricIntervalDate(sym = c("SPY","FNV","USO"), as.Date("2023-01-02"))
}
