% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{getSymPriceIntervalDate}
\alias{getSymPriceIntervalDate}
\title{getSymPriceIntervalDate}
\usage{
getSymPriceIntervalDate(
  sym,
  from_date,
  to_date = Sys.Date(),
  OHLCVA = "Adjusted"
)
}
\arguments{
\item{sym}{one or a vector of symbols}

\item{from_date}{a start date from which to retrieve symbols}

\item{to_date}{a end date to which to retrieve symbols - Default is today}

\item{OHLCVA}{column to retrieve from Yahoo service, can be any of: "Open" "High" "Low" "Close" "Volume" "Adjusted". Default value is "Adjusted".}
}
\value{
a xts matrix: each column of data contains the adjusted prices - column name is the symbol name
}
\description{
This function retrieves all values from \code{from_date} till today (or yesterday depending upon tickers)
This function calls \code{getSymFromDate} with one or a vector of tickers.
}
\examples{
getSymPriceIntervalDate(sym=c("SPY","FNV","USO"),as.Date("2023-01-02"))
}
