% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getTradeData}
\alias{getTradeData}
\title{getTradeData}
\usage{
getTradeData(TradeNr)
}
\arguments{
\item{TradeNr}{an integer or vector of integers}
}
\value{
The following fields are returned:
\itemize{
\item{TradeNr}
\item{Account: Simu/Live}
\item{TradeDate: YYYYMMDD}
\item{Strategy TBILL, OFI, WHEEL, A14,... NA also possible}
\item{Instrument}
\item{Ssjacent: a.k.a symbol}
\item{Pos: Integer}
\item{Prix: numeric}
\item{Comm.: idem}
\item{Total: ditto}
\item{Risk}
\item{Reward}
\item{PnL}
\item{Statut: Fermé, Ouvert or Ajusté}
\item{Currency: USD, EUR, CHF,...}
\item{Remarques: Text}
}
}
\description{
This function retrieves all data related to a given trade
}
\details{
It works by querying the Trades table in DB
If TradeNr does not exist then it will return a record with 0 row
}
