% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getTradeNr}
\alias{getTradeNr}
\title{getTradeNr}
\usage{
getTradeNr(v_instrument, account_type = NA, unique = T)
}
\arguments{
\item{v_instrument}{String on IBKR format, or vector of strings. Such as "SPY 15DEC23 400 P"}

\item{account_type}{one of the string values ("Live", "Simu", or NA), specifies with which account type trades are to be retrieved.
Default value is NA, which means no filtering done on account_type}

\item{unique}{Boolean - if TRUE then will return only unique trade number values otherwise will return one trade number per instrument}
}
\value{
Integer or a vector of integers
}
\description{
This function retrieves one or a vector of trade numbers, all pertaining to open/adjusted trades
including instruments listed in \code{v_instrument} argument.
}
\details{
It works by matching all instruments with Trades table in DB,
If no instruments are retrieved then NA is returned and an error message is displayed.
}
