% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullify.R
\docType{data}
\name{Global}
\alias{Global}
\title{Global variables}
\format{
An object of class \code{NULL} of length 0.
}
\description{
These variables are defined to avoid R CMD check "no visible binding" notes
when using dplyr's non-standard evaluation.
}
\details{
These global definitions remove warning messages from check().
See also https://r-pkgs.org/package-within.html
}
\keyword{internal}
