% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{greeksNet}
\alias{greeksNet}
\title{greeksNet}
\usage{
greeksNet(portf)
}
\arguments{
\item{portf}{a data frame with one line per instrument, may be grouped by date and time.
It should contain at least the following columns: \code{type;pos; mktPrice; delta; gamma; vega; theta; uPrice;
multiplier; currency} - see also readPortfolio function}
}
\value{
a data frame of double numbers with \code{delta, deltadollars, gamma, theta, vega} for each group. It is worth noticing that deltadollars is an amount in USD.
(if necessary it is converted to USD from original currency).
}
\description{
This function computes for a portfolio the net position of each Greek, summing over all positions the Greek value of each individual position.
}
\details{
Each position will be multiplied by multiplier and a Greek to obtain the Greek net value fo the position.
All Greek net values will be then summed up over all positions, for each Greek.
}
