% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_summary.R
\name{group_summary}
\alias{group_summary}
\title{A Function For Quick Access to SE & Means for Groups}
\usage{
group_summary(var, df, groupvar, autoplot = FALSE)
}
\arguments{
\item{var}{a continuous variable}

\item{df}{a dataframe containing x & y}

\item{groupvar}{a grouping variable}

\item{autoplot}{Would you like a quick visual of the output? Default value is FALSE.}
}
\description{
This function allows you to quickly find the mean & se for any variable sorted by groups. It also has an option to visualize the mean and the se quickly and easily.
}
\examples{
group_summary(Sepal.Width, iris, groupvar = Species)
group_summary(Sepal.Width, iris, groupvar = Species, autoplot = TRUE)
}
