% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{isIBAvailable}
\alias{isIBAvailable}
\title{isIBAvailable}
\usage{
isIBAvailable(silent = TRUE)
}
\arguments{
\item{silent}{Optional- print debug info, default is no info print}
}
\value{
TRUE or FALSE
}
\description{
This function tries to open a connection to TWS API on 7496 port. If successful, it returns TRUE and disconnect.
Otherwise returns FALSE
}
\examples{
\dontrun{
isIBAvailable(silent=FALSE)
}
}
