% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{isTradeOpened}
\alias{isTradeOpened}
\title{isTradeOpened}
\usage{
isTradeOpened(TradeNr)
}
\arguments{
\item{TradeNr}{an integer or vector of integers}
}
\value{
a boolean or a vector of booleans
}
\description{
This function retrieves the status of a vector of trade numers and returns a boolean vector
}
\details{
It works by querying the Trades table in DB and comparing active trades in DB with argument.
If TradeNr does not exist yet it will return FALSE just like if it were closed.
}
