% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{length.ldtv}
\alias{length.ldtv}
\title{Get Length of Data in a Variable}
\usage{
\method{length}{ldtv}(x)
}
\arguments{
\item{x}{Variable with \code{data} field.}
}
\value{
Length of \code{data} in \code{x}.
}
\description{
Get Length of Data in a Variable
}
