% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sousligne.R
\name{ligneApp}
\alias{ligneApp}
\title{ligneApp function}
\usage{
ligneApp(interest_rate)
}
\arguments{
\item{interest_rate}{this is used by caller to identify line and have the link with server piece}
}
\description{
This module server function returns expected value of a specific option in the future using BS model.
}
\details{
For computing option start price, it is possible either to retrieve current value from IBKR (NBBO option),
or to enter it manually (Manual option) or by giving a start IV so that a BS option value can be computed.

For the end value, BS will be used in all cases for computation.
Therefore all input data necessary for BS computation
-including risk free interest rate, dividend yield- are expected.
}
