% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_f.R
\name{minus.freqs}
\alias{minus.freqs}
\title{Get Interval between two frequencies}
\usage{
minus.freqs(freq1, freq2)
}
\arguments{
\item{freq1}{The first frequency.}

\item{freq2}{The second frequency.}
}
\value{
The number of intervals between the two frequencies (\code{freq1} - \code{freq2}).
}
\description{
Use this function to get the number of intervals between two frequencies.
}
\examples{
f1 <- f.yearly(2000)
f2 <- f.yearly(2010)
count <- minus.freqs(f1, f2) # this is -10
count <- minus.freqs(f2, f1) # this is 10
}
