% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{modifyJournalEntry}
\alias{modifyJournalEntry}
\title{modifyJournalEntry}
\usage{
modifyJournalEntry(entryId, close, change, text)
}
\arguments{
\item{entryId, }{integer - entry key to modify into the Journal}

\item{close, }{numeric, equal to last close value for the symbol}

\item{change, }{character - percentage change between last day and penultimate day}

\item{text, }{string - comment, remark}
}
\value{
number of lines modified, if no error it returns 1. It returns 0 if there is an error
}
\description{
This function updates an entry into Journal table.
}
\details{
It checks first that close is numeric and change is character. Otherwise it just sends the update request to DB.
}
