% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_f.R
\name{next.freq}
\alias{next.freq}
\title{Get Next Frequency}
\usage{
next.freq(freq, count)
}
\arguments{
\item{freq}{A frequency.}

\item{count}{Determines the number of steps. If negative,
it returns the previous frequency.}
}
\value{
The next frequency after the given frequency.
}
\description{
Use this function to get the next frequency.
}
\examples{
f <- f.yearly(2000)
fn <- next.freq(f, 10) # this is 2010
}
