% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{A Function to normalze data}
\usage{
normalize(x)
}
\arguments{
\item{x}{a continuous variable}
}
\description{
This function allows you to quickly normalize a vector of data.
}
\examples{
normalize(iris$Sepal.Width)
}
