% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_f.R
\name{print.ldtf}
\alias{print.ldtf}
\title{Prints an \code{ldtf} object}
\usage{
\method{print}{ldtf}(x, ...)
}
\arguments{
\item{x}{An \code{ldtf} object}

\item{...}{additional arguments}
}
\value{
\code{NULL}
}
\description{
Prints an \code{ldtf} object
}
