% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{print.ldtv}
\alias{print.ldtv}
\title{Prints a Variable}
\usage{
\method{print}{ldtv}(x, ...)
}
\arguments{
\item{x}{An \code{ldtv} object}

\item{...}{additional arguments}
}
\value{
\code{NULL}
}
\description{
Use this to print an \code{ldtv} object.
}
