% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{readAccount}
\alias{readAccount}
\title{readAccount}
\usage{
readAccount(accountnr)
}
\arguments{
\item{accountnr}{is the account number (IBKR)}
}
\value{
a tibble with the following fields: \code{ account	date	heure
NetLiquidation	EquityWithLoanValue	FullAvailableFunds	FullInitMarginReq	FullMaintMarginReq
FullExcessLiquidity	OptionMarketValue	StockMarketValue	UnrealizedPnL	RealizedPnL	TotalCashBalance
 CashFlow}
}
\description{
This function reads the Account table.
It then filters data so that it matches \code{accountnr} number
Finally it formats account data with right Date and HMS format
}
\examples{
\dontrun{
readAccount("DU5555")
}
}
