% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{readJournal}
\alias{readJournal}
\title{readJournal}
\usage{
readJournal(windowDate = NA)
}
\arguments{
\item{windowDate}{an integer, with the format YYYYmmdd. If not provided then it will return all records from current date - 300 days.}
}
\value{
a tibble with the following fields: \code{ entryId	theme	date
sym	close	change	mkt_price	mkt_change
text}
}
\description{
This function reads the Journal table.
It returns all entries from the journal starting from windowDate
}
\examples{
\dontrun{
readJournal()
}
}
