% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{readJournalMaxEntryId}
\alias{readJournalMaxEntryId}
\title{readJournalMaxEntryId}
\usage{
readJournalMaxEntryId()
}
\value{
an integer
}
\description{
This function retrieves the maximal entryId number. This is a unique key in table Journal.
}
\examples{
\dontrun{
readJournalMaxEntryId()
}
}
