% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{readLastPortfolio}
\alias{readLastPortfolio}
\title{readLastPortfolio}
\usage{
readLastPortfolio(account_type)
}
\arguments{
\item{account_type}{is a string whose value is \code{Live} or \code{Simu}}
}
\value{
a data frame with the following columns:
\code{TradeNr; date; heure; secType; symbol; expdate; strike; pos}
\code{mktPrice; optPrice; mktValue; avgCost; unPnL; IV; pvDividend}
\code{delta; gamma; vega; theta; uPrice; multiplier; currency; type; Instrument}
}
\description{
This function reads the last record of a portfolio table given as entry,
and performs a bit of data wrangling before returning a data frame with all data.
}
\details{
Data wrangling:
\enumerate{
\item formats it with right internal R Date format and heure HMS format
}
}
\examples{
\dontrun{
readLastPortfolio("DU5555")
}
}
