% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{readPortfolioDate}
\alias{readPortfolioDate}
\title{readPortfolioDate}
\usage{
readPortfolioDate(portfname, date)
}
\arguments{
\item{portfname}{is a string that is a name of a portfolio table into local DB.
local DB path is retrieved through config.yaml file}

\item{date}{is a date to filter only records for a given date.}
}
\value{
a data frame with the following columns:
\code{TradeNr; heure; symbol; expdate; strike; pos}
\code{mktPrice; optPrice; mktValue; avgCost; unPnL; IV; pvDividend}
\code{delta; gamma; vega; theta; uPrice; multiplier; currency; type; Instrument; margin}
}
\description{
This function reads a portfolio table given as entry
and performs a bit of data wrangling before returning a data frame with all data.
}
\details{
Data wrangling:
\enumerate{
\item formats it with right internal R Date format and heure HMS format
}

N.B: It will first check that portfname exists as a table in DB, and returns an error message if not
}
\examples{
\dontrun{
readPortfolio("DUxxx", Sys.Date())
}
}
