% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{row.names.ldtv}
\alias{row.names.ldtv}
\title{Get Row Names of a Variable}
\usage{
\method{row.names}{ldtv}(x)
}
\arguments{
\item{x}{Variable with \code{startFrequency} field}
}
\value{
A character string vector with frequencies of the observations as the row names.
}
\description{
Get Row Names of a Variable
}
