% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_plot.R
\name{s_plot}
\alias{s_plot}
\title{A Plotting Function With Slopes & Interactions}
\usage{
s_plot(
  x,
  y,
  df,
  groupvar = FALSE,
  interaction = FALSE,
  intplot = FALSE,
  lloc = "topleft"
)
}
\arguments{
\item{x}{a continuous variable for the x axis}

\item{y}{a continuous variable for the y axis}

\item{df}{a dataframe containing x & y}

\item{groupvar}{a grouping variable}

\item{interaction}{Would you like to see the slopes of each group? Default value is FALSE.}

\item{intplot}{Would you like to see each group plotted separately with slope and confidence intervals? Default value is FALSE.}

\item{lloc}{location of legend. Default value is "topleft". Check the ?legend documentation for other location options.}
}
\description{
This function allows you to quickly visualize the slopes and confidence intervals of your data in informative ways. It is designed for use with two continuous variables and optionally one categorical moderator.
}
\examples{
s_plot(Sepal.Width, Sepal.Length, iris)
s_plot(Sepal.Width, Sepal.Length, iris, groupvar = Species)
s_plot(Sepal.Width, Sepal.Length, iris, groupvar = Species, interaction = TRUE)
s_plot(Sepal.Width, Sepal.Length, iris, groupvar = Species, interaction = TRUE, intplot = TRUE)
}
