% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{saveTrades}
\alias{saveTrades}
\title{saveTrades}
\usage{
saveTrades(trades)
}
\arguments{
\item{trades}{data frame with the following fields:
\code{TradeNr, Account, TradeDate, Strategy, Instrument, Ssjacent, Pos, Prix,
Comm., Total, Exp.Date, Risk, Reward, PnL, Statut, Currency}}
}
\value{
No value or Error code from dbWriteTable
}
\description{
This function is used by RReporting functions and allows to update DB with modified trades
It takes as parameters a trades data frame that must have the same structure as the Trades table
}
\details{
To be sure that correct data types will be used, it converts field types to the target fields, i.e.
\itemize{
\item Integer for \code{TradeNr} and \code{Pos}
\item Real (double) for \code{Prix, Comm., Total, Risk, Reward, PnL}
}
}
