% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{A Function to Calculate the SE of a Vector}
\usage{
se(x)
}
\arguments{
\item{x}{a continuous variable}
}
\description{
This function allows you to quickly calculate the standard error of a vector.
}
\examples{
se(iris$Sepal.Width)
}
