% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sousligne.R
\name{sublineServer}
\alias{sublineServer}
\title{Ligne server}
\usage{
sublineServer(
  id,
  sym,
  mul,
  s_datetime,
  i_rate,
  u_price,
  div,
  currency,
  exchange,
  tradingClass,
  exp_dates,
  strikes
)
}
\arguments{
\item{id}{this is used by caller to identify line and have the link with server piece}

\item{sym}{symbol string - may be used to retrieved value from IBKR}

\item{mul}{multiplier - usually 100, but also 10 (ESTX50)}

\item{s_datetime}{start date time- currently now}

\item{i_rate}{value for interest rate (options BS computation)}

\item{u_price}{value for underlying spot price}

\item{div}{dividend yield used by BS option computation}

\item{currency}{string - "USD", "EUR" or "CHF"}

\item{exchange}{string usually "SMART", may be "CBOE" or "EUREX"}

\item{tradingClass}{string necessary to get option price from IBKR}

\item{exp_dates}{a vector of dates}

\item{strikes}{a vector of strikes}
}
\value{
a data frame built by \code{build_line} function
}
\description{
This module server function returns expected value of a specific option in the future using BS model.
}
\details{
For computing option start price, it is possible either to retrieve current value from IBKR (NBBO option),
or to enter it manually (Manual option) or by giving a start IV so that a BS option value can be computed.

For the end value, BS will be used in all cases for computation.
Therefore all input data necessary for BS computation
-including risk free interest rate, dividend yield- are expected.
}
