% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sousligne.R
\name{sublineUI}
\alias{sublineUI}
\title{Ligne module UI}
\usage{
sublineUI(id)
}
\arguments{
\item{id}{this is used by caller to identify line and have the link with server piece}
}
\description{
This module provides the ability to enter parameters to an option
And then to output corresponding option price by using either a computation model (BS), looking at IBKR interface or manually.
It also allows to give the option price given a number of days and an end unerlying price (BS)
}
