% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tdata-package.R
\docType{package}
\name{Tdata-package}
\alias{Tdata}
\alias{Tdata-package}
\title{Tdata: Provide basic utilities for trading IBKR account}
\description{
This package contains various R and Python functions that are to be used in cunjunction with an IBKR and other account. More specifically, it is interfaced with ib_insync Python package to get all TWS functions through the TWS API.
}
\author{
\strong{Maintainer}: Alexis Martin \email{aldohlys@gmail.com}

}
\keyword{internal}
