% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullify.R
\docType{data}
\name{TradeNr}
\alias{TradeNr}
\title{The “no visible binding” note is a peculiarity of using dplyr and unquoted variable names inside a package,
where the use of bare variable names (english and temp) looks suspicious.}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
TradeNr
}
\description{
The “no visible binding” note is a peculiarity of using dplyr and unquoted variable names inside a package,
where the use of bare variable names (english and temp) looks suspicious.
}
\keyword{datasets}
