% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_rate_utils.R
\name{update_static_yields}
\alias{update_static_yields}
\title{Update the static Swiss bond yield values in the script}
\usage{
update_static_yields(new_values, script_path = NULL)
}
\arguments{
\item{new_values}{A data frame with columns 'maturity' and 'yield'}

\item{script_path}{Path to this script file}
}
\value{
Logical indicating success
}
\description{
Update the static Swiss bond yield values in the script
}
\keyword{internal}
