% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_rate_utils.R
\name{updateDBRates}
\alias{updateDBRates}
\title{updateDBRates}
\usage{
updateDBRates(rates)
}
\arguments{
\item{rates}{Data frame containing interest rates to update, one line per interest rate. Data frame format is:
\code{Name, last_ir_update, ir1week, ir1month, ir3months, ir6months, ir1year, ir2years}.

All interest rates fields are doubles, in percent terms, and are all optional.

\code{last_ir_update} is also optional. Format is integer, YYYYMMDD. If absent it will be replaced by current date, in integer format.}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Update currency interest rates in the database
}
\details{
This function will either updates currencies interest rates on a data frame format, or one interest rate for one currency
}
\examples{
\dontrun{
updateDBRates(data.frame(Name="EUR", ir1month=2.16))
updateDBRates(data.frame(Name=c("CHF", "EUR"), ir6months = c(0.15, NA), ir1month=c(NA, 2.16)))
}
}
