% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\name{updateTicker}
\alias{updateTicker}
\title{updateTicker}
\usage{
updateTicker(name)
}
\arguments{
\item{name}{one or a vector of IBKR security names}
}
\value{
number of records updated, i.e. length(name) (normal case) or less (partial or no update, error case)
}
\description{
This function allows to update one or a vector of tickers in DB, looking at dividend yields (for STK and IND type of securities)
and at beta computations for STK, CASH and IND type of securities.
}
\examples{
\dontrun{
updateTicker("SPY")
}
}
