% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{writeJournalEntry}
\alias{writeJournalEntry}
\title{writeJournalEntry}
\usage{
writeJournalEntry(entry)
}
\arguments{
\item{entry}{a tibble, it should have the following fields: \code{ entryId    theme    date
sym    close    change    mkt_price    mkt_change}}
}
\value{
number of lines appended, if no error it returns 1. It returns 0 if there is an error
}
\description{
This function appends its argument into Journal table. No checks are performed.
}
