% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{addTicker}
\alias{addTicker}
\title{addTicker}
\usage{
addTicker(
  name,
  yahoo_name,
  trading_class,
  type = "STK",
  currency = "USD",
  exchange = "SMART",
  opt_exchange = "SMART"
)
}
\arguments{
\item{name}{IBKR security name}

\item{yahoo_name}{Yahoo finances download security name.  If not provided, will be equal to \code{name}.}

\item{trading_class}{trading class that will be used to retrieve option data from IBKR. If not provided, will be equal to \code{name}.}

\item{type}{IBKR security type, default is STK (stock). Could be also FUT (future), IND (Index),...}

\item{currency}{IBKR currency value, i.e. USD, EUR, CHF,...}

\item{exchange}{exchange name where security price can be retrieved, default is SMART}

\item{opt_exchange}{exchange name where option price for the security can be retrieved, default is SMART}
}
\value{
number of records added, i.e. 1 (normal case) or 0 (no line deleted, error case)
}
\description{
This function is used by scanner functions to add one ticker in DB
}
