% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_f.R
\name{as.character.ldtf}
\alias{as.character.ldtf}
\title{Converts Frequency to Character}
\usage{
\method{as.character}{ldtf}(x, ...)
}
\arguments{
\item{x}{value of the frequency. It must be an \code{ldtf} object returned from \code{f.?} functions.}

\item{...}{additional arguments.}
}
\value{
String representation of the value of the frequency
}
\description{
The format is explained in \code{f.?} functions.
}
