% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_f.R
\name{as.frequency}
\alias{as.frequency}
\title{Convert Character String to Frequency}
\usage{
as.frequency(str, classId)
}
\arguments{
\item{str}{The value of the frequency as a valid character, which you might have obtained from the \code{\link{as.character.ldtf}} function.}

\item{classId}{The class id of the frequency. These are explained in \code{f.?} functions.}
}
\value{
A frequency, which is an object of class 'ldtf'. See the \code{f.?} functions.
}
\description{
Use this function to convert a character string back to a frequency. You need the class id information.
}
