% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{convert_to_usd_date}
\alias{convert_to_usd_date}
\title{convert_to_usd_date}
\usage{
convert_to_usd_date(amount, currency, convert_date = Sys.Date())
}
\arguments{
\item{amount, currency}{amount is the number to be converted, currency is a string whose value is either EUR, CHF, CAD, etc..}

\item{convert_date}{Can be a date, or character, or integer(numeric). By default it is today.
If type is character/date, then \code{convert_date} argument will be first converted to an integer type with Y/M/D format
\itemize{
\item this is the format in CurrencyPairs DB table
}}
}
\description{
This function converts the amount of currency into USD, using CAD, EUR and CHF currency pairs values for a given date
}
\details{
First it loads all currency pairs that have been stored for a while, then looks up for the nearest date in the CurrencyPairs table, compared with input date.
It retrieves the EUR and CHF corresponding values.

Last it calls the convert_to_usd function.

This function can be vectorized for \code{amount} and \code{currency}, but \code{date} MUST be unique.
}
\examples{
convert_to_usd_date(100.45,"EUR",as.Date("2023-10-15"))
convert_to_usd_date(200, "CHF", 20240421)
convert_to_usd_date(200, "EUR", "20240421")
convert_to_usd_date(200, "CAD")
convert_to_usd_date(c(10000,500),c("CHF","EUR"),as.Date("2021-01-09"))
convert_to_usd_date(c(750.543,10),c("USD","EUR"),as.Date("2023-12-03"))
convert_to_usd_date(c(750.543,10),"EUR",as.Date("2023-12-03"))
}
\keyword{currency}
\keyword{trading}
