% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{currency_format}
\alias{currency_format}
\title{currency_format}
\usage{
currency_format(amount, currency)
}
\arguments{
\item{amount}{is the number to be displayed, can be also a character if convertible into a number
If length(currency) is 1, then it is recycled}

\item{currency}{a string whose value has to be defined in Currencies table from DB}
}
\description{
This function returns a string with amount and currency symbol
}
\details{
It defines a labeling functions for each currency.

Currency symbols (for EUR and USD) are also taken into consideration.

Only 2 digits after decimal are displayed. Big mark (3 digits separator) equal to space.
}
\examples{
currency_format(100.45,"EUR")
currency_format(10000,"CHF")
currency_format(1000,"CAD")
currency_format(758.458,"USD")
currency_format(100000.455,"EUR")
currency_format(c(100,40),c("EUR","USD"))
currency_format(c(100,40),"EUR")
}
