% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data.pcp}
\alias{data.pcp}
\title{Load PCP Dataset}
\usage{
data.pcp(dirPath, makeReal = FALSE)
}
\arguments{
\item{dirPath}{path to the downloaded data data.}

\item{makeReal}{uses the first column (which must be US-CPI) and converts nominal variables to real. If \code{TRUE}, \code{dirPath} must also contain a file with the US CPI.}
}
\value{
A list with data, descriptions, etc.
}
\description{
Use it to load the 'IMF's Primary Commodity Prices' dataset and create Date-Series table.
}
