% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{getAllTickers}
\alias{getAllTickers}
\title{getAllTickers}
\usage{
getAllTickers()
}
\value{
A data frame with columns \code{Name, YahooName, TradingClass, Multiplier, Type, Currency,  Exchange, OptExchange}
sorted by alphabetical order, equal to the list of current tickers stored in DB
}
\description{
This function is used by scanner functions to get all tickers to scan
}
\examples{
getAllTickers()
}
