% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{getCurrencyPairs}
\alias{getCurrencyPairs}
\title{getCurrencyPairs}
\usage{
getCurrencyPairs()
}
\description{
This function provides a pair EUR/USD, CHF/USD.
}
\details{
It first retrieves the CurrencyPairs CSV file
and looks for the last record.

If last record date is today then returns it.

Otherwise tries to connect to IBKR TWS API and retrieve pairs value.

If it does not succeed, it will then ask the end-user using enter_numerical_data  function
Once new data is obtained, then write it to CurrencyPairs file
}
\examples{
getCurrencyPairs()
}
\keyword{currency}
\keyword{trading}
