% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{getIBKRPrice}
\alias{getIBKRPrice}
\title{getIBKRPrice}
\usage{
getIBKRPrice(
  list_sec = "STK",
  list_sym,
  list_currency = "USD",
  list_exchange = "SMART",
  reqType = 2,
  close = FALSE
)
}
\arguments{
\item{list_sec}{string or vector of strings- IBKR security type, either \code{STK} or \code{IND} or \code{OPT} or \code{FUT}, default is \code{STK}.
if unknown then function returns -1}

\item{list_sym}{string or vector of strings- IBKR style of ticker, if unknown then function returns -1}

\item{list_currency}{string or vector of strings- list of currencies - either EUR, CHF or USD - default is USD}

\item{list_exchange}{string or vector of strings- list of exchanges SMART, EUREX, CBOE,..., if unknown then function returns -1 - default is SMART}

\item{reqType}{integer - should be either 2 or 4, default is 2}

\item{close}{Boolean TRUE/FALSE, default is FALSE. if TRUE then retrieve last close price from IBKR, if FALSE then retrieve market price from IBKR}
}
\value{
a data frmae with the following fields: \code{datetime} (formatted date and time), \code{sym} (ticker name), \code{price} (price as double)
}
\description{
Retrieves a price or a list of price from IBKR and
returns a data frame with date and time (formatted), the list of tickers and corresponding prices.
}
\details{
For a given ticker or a list of tickers, this function returns prices from IBKR.
If IBKR service is not available, it will return an error code and display an error message: 0 if no IBKR service, -1 if contract is unknown.
}
\examples{
\dontrun{
getIBKRPrice(list_sym="SPY")
getIBKRPrice(list_sec=c("IND","STK"), list_sym=c("ESTX50","USO"), list_currency=c("EUR","USD"),
list_exchange = c("EUREX","SMART"))
getIBKRPrice(list_sym="USO",close=TRUE)
}
}
