% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getInstrument}
\alias{getInstrument}
\title{getInstrument}
\usage{
getInstrument(account, instrument)
}
\arguments{
\item{account}{string, account name to filter into trades for the right account}

\item{instrument}{an instrument or a list of instruments as input data}
}
\value{
a data frame, and for each instrument the following fields: \code{InitialTradeDate, startPrice, DTE, u_price} and \code{IV}.
}
\description{
This function returns price, implied volatility information related to the openning trade for an instrument that is currently opened/adjuste.
}
\details{
It queries the Trades table and gets the initial trade opening date, as well as its price (without commission though).
It then retrieves interest rate at the time of opening, computes DTE and then deduce its implied volatility. It assumes dividend = 0 in IV computation.
}
\examples{
\dontrun{
getInstrument(account="Uxxx",instrument=c("CCJ 17MAY24 37 P", "USO 17MAY24 87 C"))
}
}
