% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividend_utils.R
\name{getLastDivYield}
\alias{getLastDivYield}
\title{getLastDivYield}
\usage{
getLastDivYield(names)
}
\arguments{
\item{names}{vector of tickers names - therefore all of the same type}
}
\value{
a vector of double, or NA if any missing ticker in DB or if names is not of character type
}
\description{
Looks up at DB and returns dividend yield for one or a vector of symbols.
It is expressed in decimal terms - i.e. divided by 100 after retrieved from DB
}
