% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{getLastSymPrice}
\alias{getLastSymPrice}
\title{getLastSymPrice}
\usage{
getLastSymPrice(sym)
}
\arguments{
\item{sym}{ticker name or list of tickers, as known by IBKR or Yahoo, If necessary, will be converted to Yahoo ticker name.}
}
\value{
a tibble with tidyr type of column names: \code{date, sym, value}, sym column contains the original sym list.
}
\description{
This function retrieves the last available adjusted prices of a vector of tickers from Yahoo service
}
\details{
It is based upon \code{getSymPriceIntervalDate} with from date argument equal to yesterday
}
\examples{
getLastSymPrice(c("EUR.USD","USD.CAD","CHF.USD"))
getLastSymPrice(c("SPY","XSP"))
}
