% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{getOptPrice}
\alias{getOptPrice}
\title{getOptPrice}
\usage{
getOptPrice(
  sym,
  tradingClass,
  right,
  strike,
  expiration,
  currency = "USD",
  exchange = "SMART"
)
}
\arguments{
\item{sym}{string - IBKR style of ticker, if unknown then function returns -1}

\item{tradingClass}{string - this is necessary to retrieve the correct chain for determining price - no default}

\item{right}{string - can be either P, C or Put, Call}

\item{strike}{string or numeric - option strike}

\item{expiration}{number, date or string - expiration date, format is Y/M/D}

\item{currency}{string - either EUR, CHF or USD - default is USD}

\item{exchange}{string - exchange like SMART, EUREX, CBOE,..., if unknown then function returns -1 - default is SMART}
}
\value{
a number, either option value or -1 is not found
}
\description{
Retrieves an option price from IBKR. This function is not vectorized.
}
\details{
For a given contract, this function returns a price from IBKR - it may be a last price or a mid price
If IBKR service is not available, or option price not available, or contract does not exist, it will return an error code -1.
}
\examples{
\dontrun{
getOptPrice("SPX", "SPX", "P", 5000, "20291220", "USD", "SMART")
getOptPrice("SPY", "SPY", "Call", 500.0, as.Date("2026-12-18"), "USD", "SMART")
}
}
