% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{getStockPrice}
\alias{getStockPrice}
\title{getStockPrice}
\usage{
getStockPrice(sym, close = FALSE)
}
\arguments{
\item{sym}{string - IBKR style of ticker, if unknown then function returns -1}

\item{close}{Boolean TRUE/FALSE if true then retrieve last close price if true then retrieve active price}
}
\value{
a value
}
\description{
For a given ticker this function returns the last known value from IBKR. It first looks for last value stored in prices.csv file.
If one exists that is no older than 1 hour, then this value is returned. Otherwise it looks for IBKR service to retrieve a new data.
If IBKR service is not available, it will request it from end-user on console.In the end the new price is stored in prices.csv file.
}
\details{
This function is not vectorized and accepts only one ticker at a time.
}
\examples{
\dontrun{
getStockPrice(sym="SPY")
getStockPrice(sym="ESTX50")
getStockPrice(sym="USO",close=TRUE)
}
}
