% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{getSymPrice}
\alias{getSymPrice}
\title{getSymPrice}
\usage{
getSymPrice(sym, report_date)
}
\arguments{
\item{sym}{ticker name, as known by IBKR, If necessary, will be converted to Yahoo ticker name.}

\item{report_date}{either today, then it will look at IBKR by calling getStockPrice, or a past day, then will look at Yahoo service-
\code{report_date} can be a closed day. In this case, nearest day is taken in the list.
Prices list goes back only to CurrentTradesInitialDate )set by config::get() for IBKR so currently begin of 2023
This is not suited for Gonet account}
}
\description{
This function retrieves the price of one ticker from an exchange or from Yahoo download service.
}
\details{
It tries first on Yahoo (close price) - this works only for previous days, not for today
Then in Prices table and if not available returns NA. It cannot work on a vectorized report_date
}
\examples{
getSymPrice("SPY")
}
