% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{getSymPriceFromDate}
\alias{getSymPriceFromDate}
\title{getSymPriceFromDate}
\usage{
getSymPriceFromDate(sym_list, date)
}
\arguments{
\item{sym_list}{one or a vector of symbols}

\item{date}{a start date from which to retrieve symbols}
}
\value{
a xts matrix: each column of data contains the adjusted prices - column name is the symbol name
}
\description{
This function retrieves all values from Yahoo starting from CurrentTradesInitialDate - see config.yml file at user level directory
This function calls \code{getSymFromDate} with one or a vector of tickers.
}
\examples{
getSymPriceAllDates(sym_list=c("SPY","FNV","USO"))
}
