% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{getSymPriceIntervalDate}
\alias{getSymPriceIntervalDate}
\title{getSymPriceIntervalDate}
\usage{
getSymPriceIntervalDate(sym_list, from_date, to_date = Sys.Date())
}
\arguments{
\item{sym_list}{one or a vector of symbols}

\item{from_date}{a start date from which to retrieve symbols}

\item{to_date}{a end date to which to retrieve symbols - Default is today}
}
\value{
a xts matrix: each column of data contains the adjusted prices - column name is the symbol name
}
\description{
This function retrieves all values from Yahoo starting from CurrentTradesInitialDate - see config.yml file at user level directory
This function calls \code{getSymFromDate} with one or a vector of tickers.
}
\examples{
getSymPriceIntervalDate(sym_list=c("SPY","FNV","USO"),as.Date("2023-01-02"))
}
