% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{getTickers}
\alias{getTickers}
\title{getTickers}
\usage{
getTickers()
}
\value{
A data frame with columns \code{Name, YahooName, Type, Currency, TradingClass, Multiplier, Exchange, OptExchange}
sorted by alphabetical order, equal to the list of current tickers stored in DB
}
\description{
This function is used by scanner functions to get all tickets to scan
}
