% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{greeksNet}
\alias{greeksNet}
\title{greeksNet}
\usage{
greeksNet(portf)
}
\arguments{
\item{portf}{a data frame with one line per instrument, may be grouped by date and time.
It should contain at least the following columns: \code{type;pos; mktPrice; delta; gamma; vega; theta; uPrice;
multiplier; currency} - see also readPortfolio function}
}
\value{
a data frame with \code{delta, deltadollars, gamma, theta, vega} for each group
}
\description{
This function computes net position for each Greek, using as input Greek value for each individual option
}
\details{
Each position will be multiplied by multiplier and a Greek to obtain the Greek net value.
All Greek net values will be summed over all positions, for each Greek.
}
