% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_f.R
\name{print.ldtf}
\alias{print.ldtf}
\title{Print a Frequency}
\usage{
\method{print}{ldtf}(x, ...)
}
\arguments{
\item{x}{A frequency which is the output of \code{f.?} functions in this package.}

\item{...}{Additional arguments}
}
\value{
\code{NULL}
}
\description{
Print a Frequency
}
