% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{print.ldtv}
\alias{print.ldtv}
\title{Print a Variable}
\usage{
\method{print}{ldtv}(x, ...)
}
\arguments{
\item{x}{A variable which is an object of class \code{ldtv}.}

\item{...}{Additional arguments}
}
\value{
\code{NULL}
}
\description{
Use this to print a variable.
}
