% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{readLastPortfolio}
\alias{readLastPortfolio}
\title{readLastPortfolio}
\usage{
readLastPortfolio(portfname)
}
\arguments{
\item{portfname}{is a string whose value is actual portfolio table name in DB}
}
\value{
a data frame with the following columns:
\code{TradeNr; date; heure; symbol; expdate; strike; pos; }
\code{mktPrice; optPrice; mktValue; avgCost; unPnL; IV; pvDividend; }
\code{delta; gamma; vega; theta; uPrice; multiplier; currency; type; Instrument; margin}
}
\description{
This function reads the last record of a portfolio table given as entry,
and performs a bit of data wrangling before returning a data frame with all data.
}
\details{
Data wrangling:
\enumerate{
\item format with right internal R Date format and heure HMS format: date and expdate (if it exists) fields
}

N.B: It will first check that portfname exists as a table in DB, and returns an error message if not
}
\examples{
\dontrun{
readLastPortfolio("DU5555")
}
}
