% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{readPortfolio}
\alias{readPortfolio}
\title{readPortfolio}
\usage{
readPortfolio(portfname)
}
\arguments{
\item{portfname}{is the absolute path to a portfolio file. Portfolio file is a CSV file that must follow IBKR naming rules.}
}
\value{
a tibble with the following columns:
\code{date; heure; symbol; type; expiration; strike; pos; mktPrice; optPrice}
\code{mktValue; avgCost; uPnL; IV; pvDividend; delta; gamma; vega; theta; uPrice; multiplier; currency}
}
\description{
This function reads the input file and performs a bunch of data wrangling before returning a tibble with all data.
}
\details{
Data wrangling:
\enumerate{
\item formats it with right Date and HMS format
\item converts \code{position} to \code{integer}
\item removes all CASH positions
\item rename IBKR columns to more friendly names
\item Remove secType and right columns and replace them by \code{type} column.
This assumes a right column exists in portfolio file
\code{type} value is either Stock, Put, Call or NA for anything else
}
}
\examples{
\dontrun{
readPortfolio("DU5555")
}
}
