% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{readPortfolio}
\alias{readPortfolio}
\title{readPortfolio}
\usage{
readPortfolio(portfname)
}
\arguments{
\item{portfname}{is a string that is a name of a portfolio table into local DB.
local DB path is retrieved through config.yaml file}
}
\value{
a data frame with the following columns:
\code{TradeNr; date; heure; symbol; expdate; strike; pos}
\code{mktPrice; optPrice; mktValue; avgCost; unPnL; IV; pvDividend}
\code{delta; gamma; vega; theta; uPrice; multiplier; currency; type; Instrument}
}
\description{
This function reads a portfolio table given as entry
and performs a bit of data wrangling before returning a data frame with all data.
}
\details{
Data wrangling:
\enumerate{
\item formats it with right internal R Date format and heure HMS format
}
}
\examples{
\dontrun{
readPortfolio("DU5555")
}
}
