% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{removeTicker}
\alias{removeTicker}
\title{removeTicker}
\usage{
removeTicker(name)
}
\arguments{
\item{name}{IBKR security name}
}
\value{
number of records deleted, i.e. 1 (normal case) or 0 (no line deleted, error case)
}
\description{
This function is used by scanner functions to remove one specific ticker defined by name in the DB.
}
\details{
If no name is retrieved then "Row deleted: 0" will be displayed, and 0 is returned

N.B. It could be that the same ticker name is used for different securities - I assume it is not the case in my scans
}
