% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\name{updateTicker}
\alias{updateTicker}
\title{updateTicker}
\usage{
updateTicker(name)
}
\arguments{
\item{name}{IBKR security name s}
}
\value{
number of records updated, i.e. 1 (normal case) or 0 (no line updated, error case)
}
\description{
This function is used by scanner functions to update one ticker in DB
}
\examples{
\dontrun{
updateTicker("SPY")
}
}
